+++ types/fetch-policy.d.ts
@@ -1,0 +1,1 @@
+export function fetchModel(url: any, options: any): Promise<any>;