+++ types/sass.d.ts
@@ -1,0 +1,2 @@
+declare const _default: typeof import("sass");
+export default _default;